refactor(auth): adopt feature-based structure, separate thunks from slices and update imports
