+++ frontend/src/features/auth/authActions.js
@@ -1,0 +1,30 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import * as authApi from './authService';
+
+export const loginThunk = createAsyncThunk(
+  'auth/login',
+  async (credentials, { rejectWithValue }) => {
+    try {
+      const res = await authApi.login(credentials);
+      return res.data;
+    } catch (err) {
+      return rejectWithValue(err.response.data.message);
+    }
+  },
+);
+
+export const logoutThunk = createAsyncThunk('auth/logout', async () => {
+  await authApi.logout();
+});
+
+export const signUpThunk = createAsyncThunk(
+  'auth/register',
+  async (credentials, { rejectWithValue }) => {
+    try {
+      const res = await authApi.register(credentials);
+      return res.data;
+    } catch (err) {
+      return rejectWithValue(err.response?.data?.message ?? err.message);
+    }
+  },
+);
