+++ frontend/src/features/auth/authService.js
@@ -1,4 +1,4 @@
-import api from './axios';
+import api from '@/api/axios';
 
 export const login = (credentials) => api.post('/auth/login', credentials);
 export const register = (credentials) =>
