+++ frontend/src/features/auth/forms/LoginForm.jsx
@@ -3,10 +3,10 @@ import { useForm, Controller } from 'react-hook-form';
 import { Link, useNavigate } from 'react-router-dom';
 import { zodResolver } from '@hookform/resolvers/zod';
 
-import { useAuth } from '@/hooks/useAuth.js';
-import { loginThunk } from '@/store/authSlice.js';
+import { useAuth } from '@/features/auth/hooks/useAuth.js';
+import { loginThunk } from '@/features/auth/authActions';
 
-import { loginSchema } from './loginSchema';
+import { loginSchema } from '../schema/loginSchema';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import {
