+++ frontend/src/features/auth/pages/Login.jsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/features/auth/sign-up/LoginForm';
+import LoginForm from '../forms/LoginForm';
 
 export default function Login() {
   return (
